+++ src/l10n.ts
@@ -5,11 +5,9 @@ import * as l10n from '@nextcloud/l10n';
 import dayjs from 'dayjs';
 import customParseFormat from 'dayjs/plugin/customParseFormat';
 import localizedFormat from 'dayjs/plugin/localizedFormat';
-import utc from 'dayjs/plugin/utc';
 
 dayjs.extend(customParseFormat);
 dayjs.extend(localizedFormat);
-dayjs.extend(utc);
 
 import.meta.glob('../node_modules/dayjs/locale/*.js', { eager: true });
 dayjs.locale(l10n.getLocale().toLowerCase().replace('_', '-'));